バニー系ドリンク追加パッチを修正
