+++ BunnyGarden2FixMod/Patches/PurchasableItemSelectorWrapPatch.cs
@@ -6,8 +6,6 @@ using GB;
 using GB.Bar;
 using HarmonyLib;
 using UnityEngine;
-using GB.Game.Params;
-using System.Collections.Generic;
 
 namespace BunnyGarden2FixMod.Patches;
 
@@ -245,47 +243,3 @@ public static class PurchasableItemSelectorShowArrowPatch
     }
 }
 
-
-/// <summary>
-/// <c>DrinkSetParams.ToDrinkParamList()</c> の Postfix。
-/// 選択されたドリンクセット（メニュー構成）に、バニードリンクを強制的に追加する。
-///
-/// <para>
-/// ゲームの元の仕様では、イベントなどの限定的な条件下でしか
-/// バニードリンクがリストに含まれない。
-/// 本パッチでは、UIに渡される最終的なドリンクリストに対して
-/// 直接バニー系ドリンクを追加することで、進行状況に関わらず常設化する。
-/// </para>
-/// </summary>
-
-[HarmonyPatch(typeof(DrinkSetParams), nameof(DrinkSetParams.ToDrinkParamList))]
-public class AddBunnyDrinksToDrinkSetParamsPatch
-{
-    static void Postfix(ref List<DrinkParam> __result)
-    {
-        if (__result == null) return;
-
-        // 全ドリンクリストを取得
-        List<DrinkParam> allDrinks = GBSystem.Instance.RefDrinkParams();
-
-        // 追加したいバニードリンクの ID (DrinkMenus)
-        DrinkMenus[] targetMenus = {
-            DrinkMenus.BUNNY_TRAP,
-            DrinkMenus.BUNNY_MAX,
-            DrinkMenus.BUNNY_PUNCH
-        };
-
-        foreach (var menuId in targetMenus)
-        {
-            // すでにリストに入っていないか確認
-            // allDrinks[(int)menuId] で取得したインスタンスそのものが含まれているかチェック
-            DrinkParam bunnyDrink = allDrinks[(int)menuId];
-
-            if (!__result.Contains(bunnyDrink))
-            {
-                PatchLogger.LogInfo($"Forcing {menuId} into the menu list.");
-                __result.Add(bunnyDrink);
-            }
-        }
-    }
-}
